Revisi Berita Acara Ujian

+++ application/controllers/Dashboard.php
@@ -2055,17 +2055,17 @@ class Dashboard extends CI_Controller {
 		$RekapNilai = explode("$",$Data['Mhs']['NilaiProposal1']);
 		$NilaiKetuaPenguji = 0;
 		for ($i=0; $i < count($Bobot); $i++) { 
-			$NilaiKetuaPenguji += $Bobot[$i]*(int)$RekapNilai[$i];
+			$NilaiKetuaPenguji += $Bobot[$i]*(float)$RekapNilai[$i];
 		}
 		$RekapNilai = explode("$",$Data['Mhs']['NilaiProposal2']);
 		$NilaiAnggotaPenguji = 0;
 		for ($i=0; $i < count($Bobot); $i++) { 
-			$NilaiAnggotaPenguji += $Bobot[$i]*(int)$RekapNilai[$i];
+			$NilaiAnggotaPenguji += $Bobot[$i]*(float)$RekapNilai[$i];
 		}
 		$RekapNilai = explode("$",$Data['Mhs']['NilaiProposal3']);
 		$NilaiSekretaris = 0;
 		for ($i=0; $i < count($Bobot); $i++) { 
-			$NilaiSekretaris += $Bobot[$i]*(int)$RekapNilai[$i];
+			$NilaiSekretaris += $Bobot[$i]*(float)$RekapNilai[$i];
 		}
 		$Data['Nilai'] = number_format(((0.3*$NilaiKetuaPenguji)+(0.3*$NilaiAnggotaPenguji)+(0.4*$NilaiSekretaris)),2,",",".");
 		$this->load->view('BeritaAcaraUjianProposal',$Data);
@@ -2080,17 +2080,17 @@ class Dashboard extends CI_Controller {
 		$RekapNilai = explode("$",$Data['Mhs']['NilaiProposal1']);
 		$NilaiKetuaPenguji = 0;
 		for ($i=0; $i < count($Bobot); $i++) { 
-			$NilaiKetuaPenguji += $Bobot[$i]*(int)$RekapNilai[$i];
+			$NilaiKetuaPenguji += $Bobot[$i]*(float)$RekapNilai[$i];
 		}
 		$RekapNilai = explode("$",$Data['Mhs']['NilaiProposal2']);
 		$NilaiAnggotaPenguji = 0;
 		for ($i=0; $i < count($Bobot); $i++) { 
-			$NilaiAnggotaPenguji += $Bobot[$i]*(int)$RekapNilai[$i];
+			$NilaiAnggotaPenguji += $Bobot[$i]*(float)$RekapNilai[$i];
 		}
 		$RekapNilai = explode("$",$Data['Mhs']['NilaiProposal3']);
 		$NilaiSekretaris = 0;
 		for ($i=0; $i < count($Bobot); $i++) { 
-			$NilaiSekretaris += $Bobot[$i]*(int)$RekapNilai[$i];
+			$NilaiSekretaris += $Bobot[$i]*(float)$RekapNilai[$i];
 		}
 		$Data['Nilai'] = number_format(((0.3*$NilaiKetuaPenguji)+(0.3*$NilaiAnggotaPenguji)+(0.4*$NilaiSekretaris)),2,",",".");
 		$this->load->view('ExcelUjianProposal',$Data);
@@ -2109,34 +2109,34 @@ class Dashboard extends CI_Controller {
 		$RekapNilai = explode("$",$Data['Mhs']['NilaiProposal1']);
 		$NilaiKetuaPenguji = 0;
 		for ($i=0; $i < count($Bobot); $i++) { 
-			$NilaiKetuaPenguji += $Bobot[$i]*(int)$RekapNilai[$i];
+			$NilaiKetuaPenguji += $Bobot[$i]*(float)$RekapNilai[$i];
 		}
 		$RekapNilai = explode("$",$Data['Mhs']['NilaiProposal2']);
 		$NilaiAnggotaPenguji = 0;
 		for ($i=0; $i < count($Bobot); $i++) { 
-			$NilaiAnggotaPenguji += $Bobot[$i]*(int)$RekapNilai[$i];
+			$NilaiAnggotaPenguji += $Bobot[$i]*(float)$RekapNilai[$i];
 		}
 		$RekapNilai = explode("$",$Data['Mhs']['NilaiProposal3']);
 		$NilaiSekretaris = 0;
 		for ($i=0; $i < count($Bobot); $i++) { 
-			$NilaiSekretaris += $Bobot[$i]*(int)$RekapNilai[$i];
+			$NilaiSekretaris += $Bobot[$i]*(float)$RekapNilai[$i];
 		}
 		$NilaiProposal = (0.3*$NilaiKetuaPenguji)+(0.3*$NilaiAnggotaPenguji)+(0.4*$NilaiSekretaris);
 		$Bobot = array(2.5,2.5,2,2,2,2.5,2.5,2,2.5,2.5,2);
 		$RekapNilai = explode("$",$Data['Mhs']['NilaiSkripsi1']);
 		$NilaiKetuaPenguji = 0;
 		for ($i=0; $i < count($Bobot); $i++) { 
-			$NilaiKetuaPenguji += $Bobot[$i]*(int)$RekapNilai[$i];
+			$NilaiKetuaPenguji += $Bobot[$i]*(float)$RekapNilai[$i];
 		}
 		$RekapNilai = explode("$",$Data['Mhs']['NilaiSkripsi2']);
 		$NilaiAnggotaPenguji = 0;
 		for ($i=0; $i < count($Bobot); $i++) { 
-			$NilaiAnggotaPenguji += $Bobot[$i]*(int)$RekapNilai[$i];
+			$NilaiAnggotaPenguji += $Bobot[$i]*(float)$RekapNilai[$i];
 		}
 		$RekapNilai = explode("$",$Data['Mhs']['NilaiSkripsi3']);
 		$NilaiSekretaris = 0;
 		for ($i=0; $i < count($Bobot); $i++) { 
-			$NilaiSekretaris += $Bobot[$i]*(int)$RekapNilai[$i];
+			$NilaiSekretaris += $Bobot[$i]*(float)$RekapNilai[$i];
 		}
 		$NilaiSkripsi = (0.3*$NilaiKetuaPenguji)+(0.3*$NilaiAnggotaPenguji)+(0.4*$NilaiSekretaris);
 		$Data['Nilai'] = number_format(((0.3*$NilaiProposal)+(0.7*$NilaiSkripsi)),2,",",".");
@@ -2169,17 +2169,17 @@ class Dashboard extends CI_Controller {
 		$RekapNilai = explode("$",$Data['Mhs']['NilaiProposal1']);
 		$NilaiKetuaPenguji = 0;
 		for ($i=0; $i < count($Bobot); $i++) { 
-			$NilaiKetuaPenguji += $Bobot[$i]*(int)$RekapNilai[$i];
+			$NilaiKetuaPenguji += $Bobot[$i]*(float)$RekapNilai[$i];
 		}
 		$RekapNilai = explode("$",$Data['Mhs']['NilaiProposal2']);
 		$NilaiAnggotaPenguji = 0;
 		for ($i=0; $i < count($Bobot); $i++) { 
-			$NilaiAnggotaPenguji += $Bobot[$i]*(int)$RekapNilai[$i];
+			$NilaiAnggotaPenguji += $Bobot[$i]*(float)$RekapNilai[$i];
 		}
 		$RekapNilai = explode("$",$Data['Mhs']['NilaiProposal3']);
 		$NilaiSekretaris = 0;
 		for ($i=0; $i < count($Bobot); $i++) { 
-			$NilaiSekretaris += $Bobot[$i]*(int)$RekapNilai[$i];
+			$NilaiSekretaris += $Bobot[$i]*(float)$RekapNilai[$i];
 		}
 		$NilaiProposal = (0.3*$NilaiKetuaPenguji)+(0.3*$NilaiAnggotaPenguji)+(0.4*$NilaiSekretaris);
 		$Data['NilaiProposal'] = $NilaiProposal;
@@ -2187,17 +2187,17 @@ class Dashboard extends CI_Controller {
 		$RekapNilai = explode("$",$Data['Mhs']['NilaiSkripsi1']);
 		$NilaiKetuaPenguji = 0;
 		for ($i=0; $i < count($Bobot); $i++) { 
-			$NilaiKetuaPenguji += $Bobot[$i]*(int)$RekapNilai[$i];
+			$NilaiKetuaPenguji += $Bobot[$i]*(float)$RekapNilai[$i];
 		}
 		$RekapNilai = explode("$",$Data['Mhs']['NilaiSkripsi2']);
 		$NilaiAnggotaPenguji = 0;
 		for ($i=0; $i < count($Bobot); $i++) { 
-			$NilaiAnggotaPenguji += $Bobot[$i]*(int)$RekapNilai[$i];
+			$NilaiAnggotaPenguji += $Bobot[$i]*(float)$RekapNilai[$i];
 		}
 		$RekapNilai = explode("$",$Data['Mhs']['NilaiSkripsi3']);
 		$NilaiSekretaris = 0;
 		for ($i=0; $i < count($Bobot); $i++) { 
-			$NilaiSekretaris += $Bobot[$i]*(int)$RekapNilai[$i];
+			$NilaiSekretaris += $Bobot[$i]*(float)$RekapNilai[$i];
 		}
 		$NilaiSkripsi = (0.3*$NilaiKetuaPenguji)+(0.3*$NilaiAnggotaPenguji)+(0.4*$NilaiSekretaris);
 		$Data['NilaiSkripsi'] = $NilaiSkripsi;
